expand component card's attributes with header-image and footer-image
Add the attributes "image-header" and "image-footer" to atriubue
manager, components definition list and subsequently to component
card. Including documentation and tests.

Also slip in a fix for the test-image and an adation of the
modal tests to accomadate for this.

finally, sort of deprecate emergency loading of Extension:Bootstrap.
To be removed in next major update.

closes #30